Shader追加・修正・警告削除・UI追加等

+++ DX12/Depth2.cpp
@@ -231,7 +231,7 @@ void Depth2::Draw(int TexNum1)
 		}
 		GraphVertBuff[i]->Unmap(0, nullptr);
 		//ビューポートの設定コマンド
-		DirectXBase::cmdList->RSSetViewports(1, &CD3DX12_VIEWPORT(0.0f, 0.0f, LONG(DirectXBase::Win_Width), LONG(DirectXBase::Win_Height)));
+		DirectXBase::cmdList->RSSetViewports(1, &CD3DX12_VIEWPORT(0.0f, 0.0f, DirectXBase::Win_Width, DirectXBase::Win_Height));
 		//シザー矩形の設定コマンド
 		DirectXBase::cmdList->RSSetScissorRects(1, &CD3DX12_RECT(0, 0, LONG(DirectXBase::Win_Width), LONG(DirectXBase::Win_Height)));
 		//ルートシグネチャ
